CHROMIUM: config: Disable CONFIG_RTC_SYSTOHC
This option will sync system time to RTC regularly, but only
matters if something in userspace is clearing STA_UNSYNC with
an adjtimex syscall, which does not happen on Chrome OS
(tlsdated manually syncs the RTC according to its own policy).

So we can disable the option, save a little bit of developer
confusion and code space.

cat <<EOF | tee -a \
       chromeos/config/base.config
CONFIG_RTC_SYSTOHC=n
EOF

./chromeos/scripts/kernelconfig olddefconfig

BUG=b:184702994
TEST=Boot asurada

Change-Id: Ia335499a19e6fe8f23b0df8314b46c38553c159a
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2809921
Reviewed-by: Guenter Roeck <groeck@chromium.org>