CHROMIUM: drm/amdgpu: Add GFX10.3 definitions.
These definitions are needed for the modifier patches (which support
the newer HW) but are only introduced as part of patches that depend
on a bunch more. To get clean backports without all the effort of
backporting new HW support, just backport some definitions.

Signed-off-by: Bas Nieuwenhuizen <basni@chromium.org>

BUG=b:184949073
TEST=Compile chromeos-kernel-5_4

Change-Id: I0d0b76f4d7d3b94a66b32366a586c31bb9996de3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2835723
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Bas Nieuwenhuizen <basni@chromium.org>