BACKPORT: drm/dp: Add backpointer to drm_device in drm_dp_aux
This is something that we've wanted for a while now: the ability to
actually look up the respective drm_device for a given drm_dp_aux struct.
This will also allow us to transition over to using the drm_dbg_*() helpers
for debug message printing, as we'll finally have a drm_device to reference
for doing so.

Note that there is one limitation with this - because some DP AUX adapters
exist as platform devices which are initialized independently of their
respective DRM devices, one cannot rely on drm_dp_aux->drm_dev to always be
non-NULL until drm_dp_aux_register() has been called. We make sure to point
this out in the documentation for struct drm_dp_aux.

v3:
* Add WARN_ON_ONCE() to drm_dp_aux_register() if drm_dev isn't filled out

Signed-off-by: Lyude Paul <lyude@redhat.com>
Acked-by: Thierry Reding <treding@nvidia.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210423184309.207645-4-lyude@redhat.com
Reviewed-by: Dave Airlie <airlied@redhat.com>

(cherry picked from commit 6cba3fe433415b2549c909ce72601902c8254a83)

Conflicts:
   drivers/gpu/drm/amd/amdgpu/atombios_dp.c
   drivers/gpu/drm/bridge/cadence/cdns-mhdp8546-core.c
   drivers/gpu/drm/bridge/tc358767.c
   drivers/gpu/drm/i915/display/intel_dp_aux.c
   drivers/gpu/drm/tegra/dpaux.c
   drivers/gpu/drm/xlnx/zynqmp_dp.c

Drop hunks that aren't in chromeos-5.4

BUG=b:200721581
TEST=emerge-trogdor chromeos-kernel-5_4

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ie3c074d578371b7c62e31bc8b00c5eb6725089d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3406268
Reviewed-by: Rob Clark <robdclark@chromium.org>