CHROMIUM: drm/i915: Fix crash in auto_retire
The retire logic uses the 2 lower bits of the pointer to the retire
function to store flags. However, the auto_retire function is not
guaranteed to be aligned to a multiple of 4, which causes crashes as we
jump to the wrong address.

This change fixes this by aligning the function.

BUG=b:183275734
TEST=all of dEQP.vk on dedede

Change-Id: Ie4a5805c58e941c6b5861fbc2b2c951cafa69b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2852699
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Auto-Submit: Stéphane Marchesin <marcheu@chromium.org>