CHROMIUM: ath10k: enable napi poll in threaded context.
Currently the napi poll for ath10k happens in soft irq context
which is not scheduled efficiently. Use the recent kernel feature
to run NAPI poll threaded context.
TODO: b/186906375 is to follow up on this patch with upstream
and review the approach here. Once the upstream reaches a concensus,
this patch will be replaced by the upstream version.

BUG=b:161825061
TEST=Throughput performance with napi poll in threaded context looks
     better than napi poll in soft irq context.

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I23ea0b4e4d339699fbb1a82c28b6e314d59c71ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2857879
Reviewed-by: Brian Norris <briannorris@chromium.org>