CHROMIUM: rtw88: move adaptivity mechanism to firmware
Current adaptivity mechanism is achieved in driver, by periodically
referencing the IGI value and then updating related registers.
But we find that this way may halt TX activity too long if huge
and temporary energy is detected frequently. So we move the mechanism
to firmware for immediately reacting this case to recover TX rapidly.

Signed-off-by: Chin-Yen Lee <timlee@realtek.com>

BUG=b:182193223
TEST=network_WiFi_Perf,suite:wifi_matfunc with firmware v9.9.10

Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Change-Id: I1b700edf98d99db892ebfa1d163f5e9172a12a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2897073
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>