CHROMIUM: media: platform: mtk-vcodec: vp8: emit CAPTURE frame even if decoding failed
CAPTURE frames are how errors are signaled to user-space, so we need to
emit one frame with the ERROR flag set even if decoding failed. Make
this happen by simplifying the error handling of
vdec_vp8_slice_decode().

cap_to_disp() is not always called, as it will set the ERROR flag
appropriately on the CAPTURE frame so user-space can handle the error
accordingly.

Also remove the timeout member of vdec_vp8_slice_info which was used as
a local variable of this function and was ignored anyway when calling
cap_to_disp().

BUG=b:186353920
TEST=CtsSecurityTestCases
android.security.cts.StagefrightTest#testStagefright_bug_21443020 passes
on Asurada.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Change-Id: I85aae273587f3010ee6033e92286734a631b33d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2878760
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>