FROMGIT: mac80211: assure all fragments are encrypted
Do not mix plaintext and encrypted fragments in protected Wi-Fi
networks. This fixes CVE-2020-26147.

Previously, an attacker was able to first forward a legitimate encrypted
fragment towards a victim, followed by a plaintext fragment. The
encrypted and plaintext fragment would then be reassembled. For further
details see Section 6.3 and Appendix D in the paper "Fragment and Forge:
Breaking Wi-Fi Through Frame Aggregation and Fragmentation".

Because of this change there are now two equivalent conditions in the
code to determine if a received fragment requires sequential PNs, so we
also move this test to a separate function to make the code easier to
maintain.

Cc: stable@vger.kernel.org
Signed-off-by: Mathy Vanhoef <Mathy.Vanhoef@kuleuven.be>
Link: https://lore.kernel.org/r/20210511200110.30c4394bb835.I5acfdb552cc1d20c339c262315950b3eac491397@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 965a7d72e798eb7af0aa67210e37cf7ecd1c9cad
 git://git.kernel.org/pub/scm/linux/kernel/git/jberg/mac80211.git master)

BUG=b:187853103
TEST=Verified wifi_matfunc/perf

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I4021f520992d19977cb053753299d6a93776de15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2890109
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>