CHROMIUM: power: supply: pchg: Add POWER_SUPPLY_PROP_CHARGE_TYPE
Currently, power supply devices don't have a state which indicates
a device is detected but not ready to tell charging status or
battery status. This kind of state is typically seen when a
peripheral device needs pre-charging because its battery is so
depleted that digital communication can't be established.

This patch makes pchg node set POWER_SUPPLY_CHARGE_TYPE_TRICKLE to
address the state.

Stylus not in proximity:
  POWER_SUPPLY_STATUS=Unknown
  POWER_SUPPLY_CHARGE_TYPE=N/A
  POWER_SUPPLY_CAPACITY=0

Attached:
  POWER_SUPPLY_STATUS=Charging
  POWER_SUPPLY_CHARGE_TYPE=Trickle
  POWER_SUPPLY_CAPACITY=0

After a second:
  POWER_SUPPLY_STATUS=Charging
  POWER_SUPPLY_CHARGE_TYPE=Standard
  POWER_SUPPLY_CAPACITY=72

BUG=b:189323070, b:174310207
TEST=See the description above.

Change-Id: I1d632e64574693c2b1e1ae60b3013c8710817f5b
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2924424
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>