UPSTREAM: drm/bridge: ti-sn65dsi86: fix a ternary type promotion bug
The ti_sn_aux_transfer() function returns ssize_t (signed long).  It's
supposed to return negative error codes or the number of bytes
transferred.  The "ret" variable is int and the "len" variable is
unsigned int.

The problem is that with a ternary like this, the negative int is first
type promoted to unsigned int to match "len" at this point it is a high
positive value.  Then when it is type promoted to ssize_t (s64) it
remains a high positive value instead of sign extending and becoming a
negative again.

Fix this by removing the ternary.

Fixes: b137406d9679 ("drm/bridge: ti-sn65dsi86: If refclk, DP AUX can happen w/out pre-enable")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Robert Foss <robert.foss@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/YKOGogHasIyvF8nj@mwanda
(cherry picked from commit fcb93ec3e87fe2a58b83800edad9c23bc617672b)

BUG=b:184081857, b:185508970, b:179329054, b:181806002, b:182411119
TEST=With series, EDID reading works

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I6e0a5488cc4c8b44df99011b750b119ee39d2f13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2957050
Reviewed-by: Sean Paul <seanpaul@chromium.org>