FROMGIT: arm64: dts: qcom: sc7180-trogdor: Update flash freq to match reality
This spi flash part is actually being clocked at 37.5MHz, not 25MHz,
because of the way the clk driver is rounding up the rate that is
requested to the nearest supported frequency. Let's update the frequency
here, and remove the TODO because this is the fastest frequency we're
going to be able to use here.

Reviewed-by: Douglas Anderson <dianders@chromium.org>
Cc: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/20210519054030.3217704-1-swboyd@chromium.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 729046d4f1abf341b94d39036ad33506ea9f2c7a
 https://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git arm64-for-5.14)

BUG=b:117440651
TEST=get_gbb_flags.sh works; grep qspi /sys/kernel/debug/clk_summary ->
     150MHz

Change-Id: Ie92dad19752232e0b2ce9b701b994780f83bb9f0
Disallow-Recycled-Builds: test-failures
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2906567
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>