BACKPORT: PCI: Add sysfs "removable" attribute
A PCI device is "external_facing" if it's a Root Port with the ACPI
"ExternalFacingPort" property or if it has the DT "external-facing"
property.  We consider everything downstream from such a device to
be removable by user.

We're mainly concerned with consumer platforms with user accessible
Thunderbolt ports that are vulnerable to DMA attacks, and we expect those
ports to be identified by firmware as "ExternalFacingPort". Devices in
traditional hotplug slots can technically be removed, but the expectation
is that unless the port is marked with "ExternalFacingPort", such devices
are less accessible to user / may not be removed by end user, and thus not
exposed as "removable" to userspace.

This can be used to implement userspace policies tailored for
user removable devices. Eg usage:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2591812
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2795038
(code uses such an attribute to remove external PCI devices or disable
features on them as needed by the policy desired)

Acked-by: Bjorn Helgaas <bhelgaas@google.com>
Signed-off-by: Rajat Jain <rajatja@google.com>
Link: https://lore.kernel.org/r/20210524171812.18095-2-rajatja@google.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

(cherry picked from commit c037b6c818c30b6afa11dc70018fc4a075f26028)

Conflicts:
   drivers/pci/probe.c

BUG=b:173827172
TEST=Verify the "removable" attribute on a Voxel

Change-Id: I010c9d1bbc0765033fa1414da275aaaf61803a1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3085556
Tested-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Rajat Jain <rajatja@google.com>