CHROMIUM: config: Drop AS3722 configs
This MFD is only present in tegra and samsung devices, which aren't
supported on the cros 5.4 kernel. Drop them from the common config, which
also drops them from the arm64 "flavour" because the MFD isn't used.
That seems OK given that we don't really care about these devices even
in the "generic" config.

This is generated by deleting the configs and then running

 ./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show these drivers

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I713fbcf0f33f1f1a29ccb08c0eee26a56acb76e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2923106
Reviewed-by: Douglas Anderson <dianders@chromium.org>