CHROMIUM: config: Drop TPS65090/TPS6586X configs
These mfds are only present on tegra devices, which aren't supported on
the cros 5.4 kernel. Drop them from the common config, which also drops
other drivers from the arm64 "flavour" because the MFD isn't used. That
seems OK given that we don't really care about these devices even in the
"generic" config.

This is generated by deleting the configs and then running

./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show these drivers

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I88dc8a213d8d3cdc0cec53c4e7cc653e875c9df0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2923107
Reviewed-by: Douglas Anderson <dianders@chromium.org>