CHROMIUM: config: Drop BACKLIGHT_ARCXCNN
This backlight isn't present in any DT I can find, which is confusing.
Drop it from the common config so we don't waste space on it. It looks
like an arc backlight driver, where arc is the architecture that isn't
arm so this is doubly confusing.

This is generated by deleting the config and then running

 ./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show these drivers

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I0132231563d19fea9eea6501ad79fa8e55617460
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2923109
Reviewed-by: Douglas Anderson <dianders@chromium.org>