CHROMIUM: config: Move infineon TPM to rockchip and mediatek
This config is only used on rockchip and mediatek devices (so far) in
the cros 5.4 kernel. Let's move it to those configs and the arm64
generic config so we don't waste space on other arm64 kernel platforms.

This is generated by deleting the config and moving it to the other
configs and then running

 ./chromeos/scripts/kernelconfig -F olddefconfig`

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show this driver

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ic1a59289aed583f03f2a0fa4eeda6dde204d118f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2923112
Reviewed-by: Douglas Anderson <dianders@chromium.org>