UPSTREAM: net: ipa: introduce ipa_mem_id_optional()
Introduce a new function that indicates whether a given memory
region is required for a given version of IPA hardware.  Use it to
verify that all required regions are present during initialization.

Reorder the definitions of the memory region IDs to be based on
the version in which they're first defined.  Use "+" rather than
"and above" where defining the IPA versions in which memory IDs are
used, and indicate which regions are optional (many are not).

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit d39ffb97079b3f97f4f9d7cc633c78a1f69d3264)

BUG=b:177045895
TEST=LTE functional, suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I41b55af2731ad57581e69ed6a3f962f0031fea7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3002310
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>