CHROMIUM: arm64: dts: qcom: Add an extra load switch to control PP3300_TS for homestar
In some scenes, touchpanel needs to keep VCC high, such as VT2 mode or turning off the screen through F6 shortcut key, touchpanel still works normally in these scenes.

But this panel sequence requires that power VDD must be turned off when BL_EN is turned off.
If it is not turned off, the backlight will not light up when the panel is turned on again.

BUG=b:187741215
BUG=b:189520591
TEST=emerge-strongbad chromeos-kernel-5_4

Change-Id: Iefd01fffd9012f7af6019800a8ab4084f0ba6b3e
Signed-off-by: yangcong <yangcong5@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2952549
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>