UPSTREAM: net: ipa: don't index mem data array by ID
Finally the code handles the IPA memory region array in the
configuration data without assuming it is indexed by region ID.
Get rid of the array index designators where these arrays are
initialized.  As a result, there's no more need to define an
explicitly undefined memory region ID, so get rid of that.

Change ipa_mem_find() so it no longer assumes the ipa->mem[] array
is indexed by memory region ID.  Instead, have it search the array
for the entry having the requested memory ID, and return the address
of the descriptor if found.  Otherwise return NULL.

Stop allowing memory regions to be defined with zero size and zero
canary value.  Check for this condition in ipa_mem_valid_one().
As a result, it is not necessary to check for this case in
ipa_mem_config().

Finally, there is no need for IPA_MEM_UNDEFINED to be defined any
more, so get rid of it.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit c61cfb941dcd8c0529a0c0be31bb1722feaa6082)

BUG=b:177045895
TEST=LTE functional, suspend stress test

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ifdd86a49c1da7ad33bdb3c3fa4f14d0b15a9d20a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3002522
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>