CHROMIUM: hid: Emit digitizer serial number through power_supply
(This is intended to be be upstreamed, it is pending USB
HID standardization on a new Usage in this patch stack.)

HID devices that expose a battery strength can have
associated power_supply nodes. This fills in the
SERIAL_NUMBER power_supply field if the same HID device
also has a Digitizer.Transducer Serial Number* usage,
effectively allowing that particular stylus to be
identified.

One or both of Usage(Digitizer.Transducer Serial
Number) and the proposed Usage(Digitizer.Transducer
Serial Number Second 32 Bits) will be utilized.

If the field(s) are present and non-zero, the serial number
will be either 8 or 16 hex digits.

Devices are expected to emit zero if the transducer
does not have a serial number, or the serial number
has not yet been acquired; zeros will be ignored.

Note that logical min/max (and other HID item
parameters) will be ignored for these fields.

BUG=b:184854727
TEST=Checked operation + power_supply functionality on entire patch stack; coachz, dedede

Signed-off-by: Kenneth Albanowski <kenalba@google.com>
Change-Id: Ib70ba97463c0a6c187c1dbcf9fee0781094b5b6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2954604
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>