FROMGIT: drm/panel: panel-simple: Stash DP AUX bus; allow using it for DDC
If panel-simple is instantiated as a DP AUX bus endpoint then we have
access to the DP AUX bus. Let's stash it in the panel-simple
structure, leaving it NULL for the cases where the panel is
instantiated in other ways.

If we happen to have access to the DP AUX bus and we weren't provided
the ddc-i2c-bus in some other manner, let's use the DP AUX bus for it.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210611101711.v10.6.I18e60221f6d048d14d6c50a770b15f356fa75092@changeid
(cherry picked from commit cc5a3fc041f0104d97ac61b0423ec533d7a93214
 https://github.com/freedesktop/drm-misc.git drm-misc-next)

BUG=b:184081857, b:185508970, b:179329054, b:181806002, b:182411119
TEST=With series, EDID reading works

Change-Id: I18e60221f6d048d14d6c50a770b15f356fa75092
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2900852
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>