FROMGIT: iommu/io-pgtable: Introduce map_pages() as a page table op
Mapping memory into io-pgtables follows the same semantics
that unmapping memory used to follow (i.e. a buffer will be
mapped one page block per call to the io-pgtable code). This
means that it can be optimized in the same way that unmapping
memory was, so add a map_pages() callback to the io-pgtable
ops structure, so that a range of pages of the same size
can be mapped within the same call.

Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Suggested-by: Will Deacon <will@kernel.org>
Signed-off-by: Georgi Djakov <quic_c_gdjako@quicinc.com>
Link: https://lore.kernel.org/r/1623850736-389584-4-git-send-email-quic_c_gdjako@quicinc.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit ca073b55d16a83ba7e73cd313312abc68f07f293
 https://git.kernel.org/pub/scm/linux/kernel/git/joro/iommu.git core)

BUG=b:188373249
TEST=mmc tests

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I981c7fd6955ac2bfebea7cf393a1db59846c3a39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2957428
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>