FROMGIT: iommu: Split 'addr_merge' argument to iommu_pgsize() into separate parts
The 'addr_merge' parameter to iommu_pgsize() is a fabricated address
intended to describe the alignment requirements to consider when
choosing an appropriate page size. On the iommu_map() path, this address
is the logical OR of the virtual and physical addresses.

Subsequent improvements to iommu_pgsize() will need to check the
alignment of the virtual and physical components of 'addr_merge'
independently, so pass them in as separate parameters and reconstruct
'addr_merge' locally.

No functional change.

Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Signed-off-by: Georgi Djakov <quic_c_gdjako@quicinc.com>
Reviewed-by: Lu Baolu <baolu.lu@linux.intel.com>
Link: https://lore.kernel.org/r/1623850736-389584-7-git-send-email-quic_c_gdjako@quicinc.com
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 89d5b9601f70b72f524fdb7bc6cf1b1e8e20e867
 https://git.kernel.org/pub/scm/linux/kernel/git/joro/iommu.git core)

BUG=b:188373249
TEST=mmc tests

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I7929591d78c42656166715036c6bee27351e969a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2957431
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>