FIXUP: FROMLIST: BACKPORT: media: imx208: Add imx208 camera sensor driver
0-day reports:

drivers/media/i2c/imx208.c:866:6: warning: variable 'exposure_max' set but not used

Comparison with the driver in linux-next shows a number of formatting
differences, and this variable is used in the driver in -next. Update
the version in chromeos-5.4 to match the upstream version.

BUG=b:185349909
TEST=Sanity checked basic camera functions

Signed-off-by: Guenter Roeck <groeck@chromium.org>

Change-Id: I3f5760d9a63172b81c82ecbdb71696cd8ffb1185
Disallow-Recycled-Builds: test-failures
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2975999
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Lukasz Majczak <lmajczak@google.com>