CHROMIUM: i2c: Downgrade WARN splat for "Transfer while suspended"
Now that we've landed the patch ("i2c: i2c-qcom-geni: Add shutdown
callback for i2c") (both FROMGIT and through stable tree merges) we're
getting a big warning splat at shutdown time.

Stephen has debugged this to problems in how the component model works
and is addressing it but those changes will take time.

We don't believe that the problems here are huge. Mostly the bridge
chip isn't able to talk and shutdown properly, but it never did that
before anyway. Let's downgrade the splat so we don't end up filling up
the kernel warning detector with this spam.

This patch will be reverted when b:191311358 is fixed.

BUG=b:191311358
TEST=No more splat at reboot

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ie53f649f1c07ca31bbebee7fed3fd5ea3308519f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2979823
Reviewed-by: Stephen Boyd <swboyd@chromium.org>