CHROMIUM: iommu/arm-smmu: Run usb in non-strict iommu mode on Qcom platforms
This is pretty much the same argument as in CL:2986602 and also
discussed here:
* https://lore.kernel.org/r/CAD=FV=Uf3U-kC5j_g7FB6U6_Qf2PVwKJfEsB-b+bVydW4eLymg@mail.gmail.com
* https://lore.kernel.org/r/CAD=FV=UFxZH7g8gH5+M=Fv4Y-e1bsLkNkPGJhNwhvVychcGQcQ@mail.gmail.com

It should be noted that, as far as I know, we still match or exceed
the IOMMU protection for the USB controller compared to other
platforms:
* Prior ARM platforms didn't have an IOMMU in front of USB
* Apparently AMD and Intel default to non-strict for everything except
  "untrusted" PCIe devices.

Proposing this as a CHROMIUM patch because I'm fairly sure at this
point that this is the right thing to do and this is a sane way to do
it but it still might be a while before we solve this upstream. It
doesn't make sense to keep the speed penalty while discussing the
specifics of landing this upstream.

BUG=b:162092313
TEST=dd from SD card reader 74 MB/s => 86 MB/s

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I85afed7b599d8fd84129705049ddf82c8df29983
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2991743
Reviewed-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>