CHROMIUM: iwl7000: mvm: add support for OCE scan
In case the fw supports OCE scan and one of the OCE feature flags
are set in the scan request, set the corresponding flag in the
firmware scan request.
Note that new firmware that indicates OCE support does not support
the probe deferral and suppression feature (which is optional).

Change-Id: Id79b39f650103bb9b851e31ed6a0178e81988587
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 087d111a380e315f598436a5a44cc36169df0640