CHROMIUM: rtw89: follow the AP basic rates for tx mgmt frame
By default the driver uses the 1M and 6M rate (0x0) for managemnt frames
in 2G and 5G bands respectively. But when the basic rates is
configured from the mac80211, we need to send the management frames
according to the basic rates.

This commit makes the driver use the lowest basic rates to send
the management frames.

The exception is to send probe request before connection. In this situation,
'sta' is NULL, so we send it with lowest rate instead of
vif->bss_conf.basic_rates that old value may be preserved.

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>

BUG=b:178542556,b:193705965
TEST=none

Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Change-Id: Ie4532edcafaf7f1f44993d2e8aa13b4f251a076b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3099293
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>