CHROMIUM: iwl7000: mvm: remove session protection upon station removal
When the station is removed we need to remove the session protection
that may be still running.
Note that we need also to remove the session protection upon unassoc in
case the station is kept in the AUTH state.

Change-Id: If7eba02594f20dabd22d758e1c917fbca54b2ddd
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 1a5c381d05238aef2c7a57455e5dd4c01e60f77f