FROMGIT: spi: mediatek: Fix fifo transfer
Commit 3a70dd2d0503 ("spi: mediatek: fix fifo rx mode") claims that
fifo RX mode was never handled, and adds the presumably missing code
to the FIFO transfer function. However, the claim that receive data
was not handled is incorrect. It was handled as part of interrupt
handling after the transfer was complete. The code added with the above
mentioned commit reads data from the receive FIFO before the transfer
is started, which is wrong. This results in an actual transfer error
on a Hayato Chromebook.

Remove the code trying to handle receive data before the transfer is
started to fix the problem.

Fixes: 3a70dd2d0503 ("spi: mediatek: fix fifo rx mode")
Cc: Peter Hess <peter.hess@ph-home.de>
Cc: Frank Wunderlich <frank-w@public-files.de>
Cc: Tzung-Bi Shih <tzungbi@google.com>
Cc: Hsin-Yi Wang <hsinyi@google.com>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Tested-by: Hsin-Yi Wang <hsinyi@google.com>
Tested-by: Tzung-Bi Shih <tzungbi@google.com>
Link: https://lore.kernel.org/r/20210802030023.1748777-1-linux@roeck-us.net
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 0d5c3954b35eddff0da0436c31e8d721eceb7dc2
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/spi.git for-next)

BUG=b:195090162
TEST=Hayato can boot

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: Ia6263e7e30ddd56f39f94c1ebe6b87bca2a07fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3070112
Tested-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>