CHROMIUM: iwl7000: vendor-cmd: protect geo_enabled behind VENDOR_CMD flag
When we introduced the geo_enabled flag as part of a vendor command
implementation, we didn't flag the definition and the code that sets
this flag behind the VENDOR_CMD Kconfig flag.  We only use this in the
vendor command itself, so the other references need to be behind the
Kconfig flag as well.

Change-Id: I57575819f324fe9e7642aca71906342c3f8bf2ba
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: d0bcff83ec9f55830c1082fc5064d3c11118bc16