CHROMIUM: PCI: mediatek-gen3: Disable DVFSRC voltage request
PCIe Gen3 PHY layer cannot work properly when the requested
voltage is lower than a specific level(e.g. 0.55V, it depends
on the chip manufacturing process).

If DVFSRC is implemented:
The requested voltage may reduce to lower level when enter
L1 substate. When exiting L1 substate, the MAC layer asserts
a HW signal to request DVFSRC to raise voltage to normal mode.
It waits for the voltage ready signal from DVFSRC to determine
if the LTSSM can proceed.

If DVFSRC is not implemented:
The MAC layer still asserts the voltage request to DVFSRC when
exiting L1 substate. However, it will not receive the voltage
ready signal.

DVFSRC is not ready. Disables DVFSRC voltage request in case
it failed to resume.

BUG=b:188387920
TEST=emerge-asurada chromeos-kernel-5_4

Signed-off-by: cheng qizhong <qizhong.cheng@mediatek.com>
Change-Id: I485401e35537a02d72c0e55d083755dddf173df0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3080440
Tested-by: Qizhong Cheng <qizhong.cheng@mediatek.corp-partner.google.com>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>